change type variable in unittest from int to long

+++ unittest/src/test/java/FizzBuzzTest.java
@@ -7,7 +7,7 @@ public class FizzBuzzTest {
 
     @Test
     public void input_1_print_1() throws Exception {
-        int input = 1;
+        Long input = Long.valueOf(1);
         FizzBuzz fizzbuzz = new FizzBuzz();
         Assert.assertEquals("1", fizzbuzz.show(input));
 
@@ -15,7 +15,7 @@ public class FizzBuzzTest {
 
     @Test
     public void input_3_print_Fizz() throws Exception {
-        int input = 3;
+        Long input = Long.valueOf(3);
         FizzBuzz fizzbuzz = new FizzBuzz();
         Assert.assertEquals("Fizz", fizzbuzz.show(input));
 
@@ -23,7 +23,7 @@ public class FizzBuzzTest {
 
     @Test
     public void input_5_print_Buzz() throws Exception {
-        int input = 5;
+        Long input = Long.valueOf(5);
         FizzBuzz fizzbuzz = new FizzBuzz();
         Assert.assertEquals("Buzz", fizzbuzz.show(input));
 
@@ -31,9 +31,10 @@ public class FizzBuzzTest {
 
     @Test
     public void input_15_print_FizzBuzz() throws Exception {
-        int input = 15;
+        Long input = Long.valueOf(15);
         FizzBuzz fizzbuzz = new FizzBuzz();
         Assert.assertEquals("FizzBuzz", fizzbuzz.show(input));
 
     }
+
 }
